added retention on new tables

+++ source/Infrastructure/InboxEvents/RemoveMonthOldReceivedInboxEventIdsWhenADaysHasPassed.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
 using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,28 +20,32 @@ using Application.Configuration.DataAccess;
 using Application.Configuration.TimeEvents;
 using MediatR;
 using Microsoft.Data.SqlClient;
+using NodaTime;
 
-namespace Infrastructure.Configuration.IntegrationEvents;
+namespace Infrastructure.InboxEvents;
 
-public class RemoveIntegrationEventsWhenADaysHasPassed : INotificationHandler<ADayHasPassed>
+public class RemoveMonthOldReceivedInboxEventIdsWhenADaysHasPassed : INotificationHandler<ADayHasPassed>
 {
     private readonly IDatabaseConnectionFactory _databaseConnectionFactory;
 
-    public RemoveIntegrationEventsWhenADaysHasPassed(IDatabaseConnectionFactory databaseConnectionFactory)
+    public RemoveMonthOldReceivedInboxEventIdsWhenADaysHasPassed(
+        IDatabaseConnectionFactory databaseConnectionFactory)
     {
         _databaseConnectionFactory = databaseConnectionFactory;
     }
 
     public async Task Handle(ADayHasPassed notification, CancellationToken cancellationToken)
     {
-        while (await AnyReceivedIntegrationEventsAsync(cancellationToken).ConfigureAwait(false))
+        if (notification == null) throw new ArgumentNullException(nameof(notification));
+
+        while (await AnyReceivedInboxEventsAsync(notification, cancellationToken).ConfigureAwait(false))
         {
             const string deleteStmt = @"
                 WITH CTE AS
                  (
                      SELECT TOP 10000 *
-                     FROM [dbo].[ReceivedIntegrationEvents]
-                     WHERE [ProcessedDate] IS NOT NULL AND [ErrorMessage] IS NULL
+                     FROM [dbo].[ReceivedInboxEventIds]
+                     WHERE [OccurredOn] < @LastMonthInstant
                  )
                 DELETE FROM CTE;";
 
@@ -49,6 +54,9 @@ public class RemoveIntegrationEventsWhenADaysHasPassed : INotificationHandler<AD
                     .ConfigureAwait(false);
             using var transaction = connection.BeginTransaction();
             using var command = connection.CreateCommand();
+            command.Parameters.AddWithValue(
+                "@LastMonthInstant",
+                notification.Now.Plus(-Duration.FromDays(30)).ToDateTimeUtc());
             command.Transaction = transaction;
             command.CommandText = deleteStmt;
 
@@ -66,18 +74,22 @@ public class RemoveIntegrationEventsWhenADaysHasPassed : INotificationHandler<AD
         }
     }
 
-    private async Task<bool> AnyReceivedIntegrationEventsAsync(CancellationToken cancellationToken)
+    private async Task<bool> AnyReceivedInboxEventsAsync(ADayHasPassed today, CancellationToken cancellationToken)
     {
         const string selectStmt = @"
-               SELECT Count(*) FROM [dbo].[ReceivedIntegrationEvents] WHERE [ProcessedDate] IS NOT NULL AND [ErrorMessage] IS NULL";
+               SELECT Count(*) FROM [dbo].[ReceivedInboxEventIds] WHERE [OccurredOn] < @LastMonthInstant";
 
         using var connection =
-            (SqlConnection)await _databaseConnectionFactory.GetConnectionAndOpenAsync(cancellationToken).ConfigureAwait(false);
+            (SqlConnection)await _databaseConnectionFactory
+                .GetConnectionAndOpenAsync(cancellationToken).ConfigureAwait(false);
         using var command = connection.CreateCommand();
+        command.Parameters.AddWithValue(
+            "@LastMonthInstant",
+            today.Now.Plus(-Duration.FromDays(30)).ToDateTimeUtc());
         command.CommandText = selectStmt;
 
-        var amountOfProcessedEvents = (int)await command.ExecuteScalarAsync(cancellationToken).ConfigureAwait(false);
+        var amountOfOldEvents = (int)await command.ExecuteScalarAsync(cancellationToken).ConfigureAwait(false);
 
-        return amountOfProcessedEvents > 0;
+        return amountOfOldEvents > 0;
     }
 }
